Move `ContainerWidget` to its own file
